requirements.txt
streamlit
pandas
scikit-learn
Pillow
matplotlib
plotly
scipy
bokeh>=3.1.0
streamlit-bokeh
numpy<2.0,>=1.24.4